renamed the main.py file
